Move ReferenceCounter
